requirements.txt
streamlit
pillow
pandas
google-cloud-storage
google-cloud-firestore
google-auth
